refactor(graph): split 970-line graph.ts into focused modules
Split into: types.ts, parser.ts, analyzer.ts, render.ts, server.ts,
with thin index.ts orchestrator. Updated test imports.
