+++ src/commands/graph.test.ts
@@ -2,7 +2,8 @@
  * Tests for graph command --days filtering and bar chart data
  */
 import { describe, it, expect } from "vitest";
-import { calculateCutoffMs, filterByDays, analyzeSession } from "./graph.js";
+import { calculateCutoffMs, filterByDays } from "./graph/parser.js";
+import { analyzeSession } from "./graph/analyzer.js";
 
 describe("graph --days filtering", () => {
   describe("calculateCutoffMs", () => {
